Made all changes suggested in PR feedback
