+++ src/main/kotlin/com/example/library/domain/BookStatus.kt
@@ -1,0 +1,11 @@
+package com.example.library.domain
+
+enum class BookStatus {
+    AVAILABLE,
+    CHECKED_OUT;
+
+    companion object {
+        fun fromBoolean(available: Boolean): BookStatus =
+            if (available) AVAILABLE else CHECKED_OUT
+    }
+}
