+++ src/main/kotlin/com/example/library/domain/Requests.kt
@@ -1,0 +1,9 @@
+package com.example.library.domain
+
+data class AddBookRequest(val title: String)
+
+data class RegisterUserRequest(val name: String)
+
+data class CheckoutRequest(val userId: Int, val bookId: Int)
+
+data class ReturnRequest(val userId: Int, val bookId: Int)
